BR-17 - Refactoring (nothing useful)
