+++ src/test/java/com/bobocode/svydovets/core/ApplicationContextTest.java
@@ -1,9 +1,9 @@
 package com.bobocode.svydovets.core;
 
 import com.bobocode.svydovets.config.BeanConfigBase;
-import com.bobocode.svydovets.service.NonPrimaryProductServiceImpl;
-import com.bobocode.svydovets.service.PrimaryProductServiceImpl;
-import com.bobocode.svydovets.service.ProductService;
+import com.bobocode.svydovets.service.qualifier.NonPrimaryProductServiceImpl;
+import com.bobocode.svydovets.service.qualifier.PrimaryProductServiceImpl;
+import com.bobocode.svydovets.service.qualifier.ProductService;
 import com.bobocode.svydovets.service.TrimService;
 import com.bobocode.svydovets.service.base.CollectionsHolderService;
 import com.bobocode.svydovets.service.base.CommonService;
@@ -60,7 +60,6 @@ public class ApplicationContextTest {
 
     @Test
     @Order(2)
-
     void createApplicationContextFromConfigClass() {
         ApplicationContext context = new DefaultApplicationContext(BeanConfigBase.class);
         assertThat(context).isNotNull();
@@ -68,7 +67,6 @@ public class ApplicationContextTest {
 
     @Test
     @Order(3)
-
     void applicationContextFromBasePackageCreatesAllRequiredBeans() {
         String basePackage = "com.bobocode.svydovets.service.base";
         ApplicationContext context = new DefaultApplicationContext(basePackage);
@@ -83,7 +81,6 @@ public class ApplicationContextTest {
 
     @Test
     @Order(4)
-
     void applicationContextFromConfigClassCreatesAllRequiredBeans() {
         ApplicationContext context = new DefaultApplicationContext(BeanConfigBase.class);
 
@@ -98,11 +95,10 @@ public class ApplicationContextTest {
 
     @Test
     @Order(5)
-
     void getBeanThrowNoSuchBeanException() {
         String basePackage = "com.bobocode.svydovets.service.base";
         ApplicationContext context = new DefaultApplicationContext(basePackage);
-        when(packageScanner.findAllBeanByBasePackage(basePackage))
+        when(packageScanner.findComponentsByBasePackage(basePackage))
                 .thenReturn(Set.of(CommonService.class, EditService.class, MessageService.class));
         assertThatExceptionOfType(NoSuchBeanException.class)
                 .isThrownBy(() -> context.getBean(TrimService.class))
@@ -111,11 +107,10 @@ public class ApplicationContextTest {
 
     @Test
     @Order(6)
-
     void getBeanThrowNoUniqueBeanException() {
         String basePackage = "com.bobocode.svydovets.service";
         ApplicationContext context = new DefaultApplicationContext(basePackage);
-        when(packageScanner.findAllBeanByBasePackage(basePackage))
+        when(packageScanner.findComponentsByBasePackage(basePackage))
                 .thenReturn(Set.of(PrimaryProductServiceImpl.class, NonPrimaryProductServiceImpl.class));
         assertThatExceptionOfType(NoUniqueBeanException.class)
                 .isThrownBy(() -> context.getBean(ProductService.class))
@@ -124,9 +119,8 @@ public class ApplicationContextTest {
 
     @Test
     @Order(7)
-
     void injectBeansToTheList() {
-        ApplicationContext context = new DefaultApplicationContext("com.bobocode.svydovets.service");
+        ApplicationContext context = new DefaultApplicationContext("com.bobocode.svydovets.service.base");
 
         CollectionsHolderService.ListHolderService listHolderService = context.getBean(CollectionsHolderService.ListHolderService.class);
         var productServiceList = listHolderService.getProductServiceList();
@@ -144,9 +138,8 @@ public class ApplicationContextTest {
 
     @Test
     @Order(8)
-
     void injectBeansToTheSet() {
-        ApplicationContext context = new DefaultApplicationContext("com.bobocode.svydovets.service");
+        ApplicationContext context = new DefaultApplicationContext("com.bobocode.svydovets.service.base");
 
         CollectionsHolderService.SetHolderService setHolderService = context.getBean(CollectionsHolderService.SetHolderService.class);
         var productServiceList = setHolderService.getProductServiceSet();
