--- src/test/java/com/bobocode/svydovets/service/ProductService.java
@@ -1,4 +1,0 @@
-package com.bobocode.svydovets.service;
-
-public interface ProductService {
-}
