+++ src/test/java/com/bobocode/svydovets/service/qualifier/PrimaryProductServiceImpl.java
@@ -1,4 +1,4 @@
-package com.bobocode.svydovets.service;
+package com.bobocode.svydovets.service.qualifier;
 
 import svydovets.core.annotation.Component;
 import svydovets.core.annotation.Primary;
